requirements.txt
alembic==1.9.1
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
clickhouse-driver==0.2.5
clickhouse-sqlalchemy==0.2.3
greenlet==2.0.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
pycodestyle==2.10.0
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
SQLAlchemy==1.4.45
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
